Use Lists instead of Tuples
I was originally convinced to use Tuples to track my collection of LEDs, because Tuples allow for fast access to arbitrary elements.

However, this came with a tax of needing to convert to lists constantly, as well as the inability to iterate over a Tuple. Since a list is better suited for most operations, I’ve converted the collection to a list. Benchmarks show roughly equivalent performance for my use-case.